Add pieces functionality for custom sets.
