+++ LegoCollectionManager/LegoCollectionManager/Models/Colour.cs
@@ -9,6 +9,7 @@ namespace LegoCollectionManager.Models
     {
         public Colour()
         {
+            CustomSetPieces = new HashSet<CustomSetPiece>();
             MissingPieces = new HashSet<MissingPiece>();
             SetPieces = new HashSet<SetPiece>();
         }
@@ -16,6 +17,7 @@ namespace LegoCollectionManager.Models
         public int ColourId { get; set; }
         public string ColourName { get; set; }
 
+        public virtual ICollection<CustomSetPiece> CustomSetPieces { get; set; }
         public virtual ICollection<MissingPiece> MissingPieces { get; set; }
         public virtual ICollection<SetPiece> SetPieces { get; set; }
     }
