+++ LegoCollectionManager/LegoCollectionManager/Models/CustomSet.cs
@@ -7,6 +7,11 @@ namespace LegoCollectionManager.Models
 {
     public partial class CustomSet
     {
+        public CustomSet()
+        {
+            CustomSetPieces = new HashSet<CustomSetPiece>();
+        }
+
         public int CustomSetId { get; set; }
         public string CustomSetName { get; set; }
         public int? PiecesAmount { get; set; }
@@ -14,5 +19,6 @@ namespace LegoCollectionManager.Models
         public int? User { get; set; }
 
         public virtual User UserNavigation { get; set; }
+        public virtual ICollection<CustomSetPiece> CustomSetPieces { get; set; }
     }
 }
