+++ LegoCollectionManager/LegoCollectionManager/Views/CustomSet/AddPiece.cshtml
@@ -1,24 +1,27 @@
-﻿@model LegoCollectionManager.Models.SetPiece
+﻿@model LegoCollectionManager.Models.CustomSetPiece
 
-<h4>Add Piece</h4>
+<h4>CustomSetPiece</h4>
 <hr />
 <div class="row">
     <div class="col-md-4">
         <form asp-action="AddPiece">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
+                <label asp-for="CustomSet" class="control-label"></label>
+                <input asp-for="CustomSet" class ="form-control" value=@ViewData["CustomSetId"]/>
+            </div>
+            <div class="form-group">
                 <label asp-for="Piece" class="control-label"></label>
-                <select asp-for="Piece" class="form-control" asp-items="ViewBag.Pieces"/>
-                <span asp-validation-for="Piece" class="text-danger"></span>
+                <select asp-for="Piece" class ="form-control" asp-items="ViewBag.Pieces"></select>
+            </div>
+            <div class="form-group">
+                <label asp-for="Colour" class="control-label"></label>
+                <select asp-for="Colour" class ="form-control" asp-items="ViewBag.Colours"></select>
             </div>
             <div class="form-group">
                 <label asp-for="Amount" class="control-label"></label>
                 <input asp-for="Amount" class="form-control" />
                 <span asp-validation-for="Amount" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Colour" class="control-label"></label>
-                <select asp-for="Colour" class ="form-control" asp-items="ViewBag.Colours"></select>
             </div>
             <div class="form-group">
                 <input type="submit" value="Create" class="btn btn-primary" />
