+++ LegoCollectionManager/LegoCollectionManager/Views/User/AddAvatar.cshtml
@@ -17,8 +17,6 @@
                 {
                     <input name="Avatar" id=@avatar.Value type="radio" value=@avatar.Value />
                     <img src=@avatar.Text style="width: 150px"/>
-                    <h1>@avatar.Value</h1>
-                    <br />
     
                 }
             </div>
